+++ packages/backend/src/database/entities/spaces.ts
@@ -7,6 +7,7 @@ import {
 } from '@lightdash/common';
 import { Knex } from 'knex';
 import database from '../database';
+import { AnalyticsChartViewsTableName } from './analytics';
 import {
     DbPinnedList,
     PinnedChartTableName,
@@ -15,6 +16,7 @@ import {
 } from './pinnedList';
 import { ProjectTableName } from './projects';
 import { SavedChartsTableName } from './savedCharts';
+import { UserTableName } from './users';
 
 export type DbSpace = {
     space_id: number;
@@ -49,11 +51,12 @@ export type SpaceShareTable = Knex.CompositeTableType<
 
 export const SpaceShareTableName = 'space_share';
 
-export const getSpace = async (
+export const getFirstAccessibleSpace = async (
     db: Knex,
     projectUuid: string,
+    userUuid: string,
 ): Promise<DbSpace & Pick<DbPinnedList, 'pinned_list_uuid'>> => {
-    const results = await db('spaces')
+    const space = await db('spaces')
         .innerJoin('projects', 'projects.project_id', 'spaces.project_id')
         .innerJoin(
             'organizations',
@@ -70,6 +73,22 @@ export const getSpace = async (
             `${PinnedListTableName}.pinned_list_uuid`,
             `${PinnedSpaceTableName}.pinned_list_uuid`,
         )
+        .leftJoin(
+            SpaceShareTableName,
+            `${SpaceShareTableName}.space_id`,
+            `${SpaceTableName}.space_id`,
+        )
+        .leftJoin(
+            'users',
+            `${SpaceShareTableName}.user_id`,
+            `${UserTableName}.user_id`,
+        )
+        .where((q) => {
+            q.where(`${UserTableName}.user_uuid`, userUuid).orWhere(
+                `${SpaceTableName}.is_private`,
+                false,
+            );
+        })
         .where(`${ProjectTableName}.project_uuid`, projectUuid)
         .select<(DbSpace & Pick<DbPinnedList, 'pinned_list_uuid'>)[]>([
             'spaces.space_id',
@@ -80,20 +99,26 @@ export const getSpace = async (
             'organizations.organization_uuid',
             `${PinnedListTableName}.pinned_list_uuid`,
         ])
-        .limit(1);
-    const [space] = results;
+        .first();
+
     if (space === undefined) {
         throw new NotFoundError(
             `No space found for project with id: ${projectUuid}`,
         );
     }
+
     return space;
 };
 
 export const getSpaceWithQueries = async (
     projectUuid: string,
+    userUuid: string,
 ): Promise<Space> => {
-    const space = await getSpace(database, projectUuid);
+    const space = await getFirstAccessibleSpace(
+        database,
+        projectUuid,
+        userUuid,
+    );
     const savedQueries = await database('saved_queries')
         .leftJoin(
             'saved_queries_versions',
@@ -128,6 +153,7 @@ export const getSpaceWithQueries = async (
                 chart_config: ChartConfig['config'];
                 chart_type: ChartType;
                 views: string;
+                first_viewed_at: Date | null;
             }[]
         >([
             `saved_queries.saved_query_uuid`,
@@ -142,7 +168,10 @@ export const getSpaceWithQueries = async (
             `${PinnedListTableName}.pinned_list_uuid`,
 
             database.raw(
-                `(SELECT COUNT('analytics_chart_views.chart_uuid') FROM analytics_chart_views WHERE saved_queries.saved_query_uuid = analytics_chart_views.chart_uuid) as views`,
+                `(SELECT COUNT('${AnalyticsChartViewsTableName}.chart_uuid') FROM ${AnalyticsChartViewsTableName} WHERE saved_queries.saved_query_uuid = ${AnalyticsChartViewsTableName}.chart_uuid) as views`,
+            ),
+            database.raw(
+                `(SELECT ${AnalyticsChartViewsTableName}.timestamp FROM ${AnalyticsChartViewsTableName} WHERE saved_queries.saved_query_uuid = ${AnalyticsChartViewsTableName}.chart_uuid ORDER BY ${AnalyticsChartViewsTableName}.timestamp ASC LIMIT 1) as first_viewed_at`,
             ),
         ])
         .orderBy([
@@ -180,6 +209,7 @@ export const getSpaceWithQueries = async (
                 savedQuery.chart_config,
             ),
             views: parseInt(savedQuery.views, 10) || 0,
+            firstViewedAt: savedQuery.first_viewed_at,
         })),
         projectUuid,
         dashboards: [],