+++ packages/backend/src/services/UserService.ts
@@ -26,6 +26,7 @@ import {
     SessionUser,
     UpdateUserArgs,
     UserAllowedOrganization,
+    validateOrganizationEmailDomains,
 } from '@lightdash/common';
 import { randomInt } from 'crypto';
 import { nanoid } from 'nanoid';
@@ -421,10 +422,16 @@ export class UserService {
                 },
             });
             if (enableEmailDomainAccess && user.email) {
+                const emailDomain = getEmailDomain(user.email);
+
+                const error = validateOrganizationEmailDomains([emailDomain]);
+                if (error) {
+                    throw new ParameterError(error);
+                }
                 await this.organizationAllowedEmailDomainsModel.upsertAllowedEmailDomains(
                     {
                         organizationUuid: user.organizationUuid,
-                        emailDomains: [getEmailDomain(user.email)],
+                        emailDomains: [emailDomain],
                         role: OrganizationMemberRole.VIEWER,
                         projectUuids: [],
                     },
@@ -687,7 +694,11 @@ export class UserService {
     async sendOneTimePasscodeToPrimaryEmail(
         user: Pick<SessionUser, 'userUuid'>,
     ): Promise<EmailStatusExpiring> {
-        const passcode = randomInt(999999).toString().padStart(6, '0');
+        const passcode =
+            lightdashConfig.mode === LightdashMode.PR ||
+            lightdashConfig.mode === LightdashMode.DEV
+                ? '000000'
+                : randomInt(999999).toString().padStart(6, '0');
         const emailStatus = await this.emailModel.createPrimaryEmailOtp({
             passcode,
             userUuid: user.userUuid,